printf warning

+++ src/lgfx/v1_autodetect/LGFX_AutoDetect_ESP32_all.hpp
@@ -3275,7 +3275,7 @@ namespace lgfx
 
 
       std::uint32_t pkg_ver = lgfx::get_pkg_ver();
-      ESP_LOGV("LGFX", "pkg: %d", pkg_ver);
+      ESP_LOGV("LGFX", "pkg: %lu", (unsigned long)pkg_ver);
 
       switch (pkg_ver)
       {
